Rename javascript files to follow conventions * See https://github.com/airbnb/javascript#naming-conventions
