+++ client/app/api/Base.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { csrfToken } from 'lib/helpers/serverContext';
+import { csrfToken } from 'lib/helpers/server-context';
 
 export default class BaseAPI {
   constructor() {
