+++ client/app/bundles/course/enrol_requests.js
@@ -1,3 +1,3 @@
-import initializeAjaxForms from 'lib/helpers/tableRowFormHelpers';
+import initializeAjaxForms from 'lib/helpers/initializeAjaxForms';
 
 initializeAjaxForms('tr.course-enrol-request .approve');
