+++ client/app/bundles/course/experience_points_records.js
@@ -1,3 +1,3 @@
-import initializeAjaxForms from 'lib/helpers/tableRowFormHelpers';
+import initializeAjaxForms from 'lib/helpers/initializeAjaxForms';
 
 initializeAjaxForms('tr.experience-points-record #update');
