+++ client/app/bundles/course/users.js
@@ -1,3 +1,3 @@
-import initializeAjaxForms from 'lib/helpers/tableRowFormHelpers';
+import initializeAjaxForms from 'lib/helpers/initializeAjaxForms';
 
 initializeAjaxForms('tr.course-user #update');
