+++ client/app/bundles/system/admin/instance/users.js
@@ -1,3 +1,3 @@
-import initializeAjaxForms from 'lib/helpers/tableRowFormHelpers';
+import initializeAjaxForms from 'lib/helpers/initializeAjaxForms';
 
 initializeAjaxForms('tr.instance-user #update');
