+++ client/app/bundles/system/admin/users.js
@@ -1,3 +1,3 @@
-import initializeAjaxForms from 'lib/helpers/tableRowFormHelpers';
+import initializeAjaxForms from 'lib/helpers/initializeAjaxForms';
 
 initializeAjaxForms('tr.user #update');
