+++ client/app/lib/helpers/ace-editor.js
