+++ client/app/lib/helpers/rails-helpers.js
