+++ client/app/lib/helpers/server-context.js
