+++ client/app/lib/helpers/url-helpers.js
@@ -6,8 +6,7 @@
 *
 * Taken from http://stackoverflow.com/questions/19491336/get-url-parameter-jquery-or-how-to-get-query-string-values-in-js
 */
-
-export default function getUrlParameter(sParam) {
+function getUrlParameter(sParam) {
   const sPageURL = decodeURIComponent(window.location.search.substring(1));
   const sURLVariables = sPageURL.split('&');
   let sParameterName;
@@ -21,3 +20,6 @@ export default function getUrlParameter(sParam) {
   }
   return '';
 }
+
+/* eslint-disable import/prefer-default-export */
+export { getUrlParameter };
