+++ client/app/lib/moment.js
@@ -1,5 +1,5 @@
 import moment from 'moment-timezone';
-import { timeZone } from 'lib/helpers/serverContext';
+import { timeZone } from 'lib/helpers/server-context';
 
 moment.tz.setDefault(timeZone);
 export default moment;
